fix: `finfo_object`, `iinfo_object`, `_array` to typing.Protocol
